add DESIGN.md (#1)
* add DESIGN.md

This is intended as an initial space where we can discuss the design of
`log-monster` before getting too deep into implementation.  The content here
is not intended to be "final" and may change as we explore and discuss new
design ideas.

Eventually, this document will likely be removed in favor of slightly more
formal API documentation and design documents.

* fix typo

* fix more typos